10/more queries

+++ server/db/migrations/20170408025446-create-provider.js
@@ -18,7 +18,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       phone: {
-        type: Sequelize.INTEGER
+        type: Sequelize.BIGINT(11)
       },
       createdAt: {
         allowNull: false,
